requirements.txt
asgiref==3.2.3
attrdict==2.0.1
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cmake==3.16.3
cymem==2.0.3
Django==3.0.3
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
joblib==0.14.1
mojimoji==0.0.9
murmurhash==1.0.2
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
pyknp==0.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
sentencepiece==0.1.85
six==1.14.0
spacy==2.2.3
sqlparse==0.3.0
srsly==1.0.1
thinc==7.3.1
torch==1.4.0
torchtext==0.5.0
torchvision==0.5.0
tqdm==4.43.0
urllib3==1.25.8
wasabi==0.6.0
zipp==3.0.0
